+++ src/jtl/Connector/Shopware/Controller/Product.php
@@ -145,7 +145,6 @@ class Product extends DataController
 
         // ProductPrice
         $recommendedRetailPrice = 0.0;
-        $purchasePrice = 0.0;
         $customerGroupCache = null;
         $productPriceId = new Identity(IdConcatenator::link(array($product->getId()->getEndpoint(), 0)));
         $productPrice = Mmc::getModel('ProductPrice');
@@ -193,7 +192,6 @@ class Product extends DataController
             // Search default product price
             if ($customerGroup->getId() == Shopware()->Shop()->getCustomerGroup()->getId() && (int) $data['prices'][$i]['from'] == 1) {
                 $recommendedRetailPrice = (double) $data['prices'][$i]['pseudoPrice'];
-                $purchasePrice = (double) $data['prices'][$i]['basePrice'];
                 $defaultPrice = clone $productPrice;
                 $defaultPrice->setCustomerGroupId(new Identity('0', 0))
                     ->setCustomerId(new Identity('0', 0));
@@ -210,8 +208,7 @@ class Product extends DataController
         // add default price
         if ($defaultPrice !== null) {
             $product->addPrice($defaultPrice)
-                ->setRecommendedRetailPrice($recommendedRetailPrice)
-                ->setPurchasePrice($purchasePrice);
+                ->setRecommendedRetailPrice($recommendedRetailPrice);
         } else {
             Logger::write(sprintf('Could not find any default price for product (%s, %s)', 
                 $product->getId()->getEndpoint(),